+++ tests/testapp/tests/sync/test_controller.py
@@ -1,15 +1,16 @@
+import contextlib
 import json
 import uuid
-import contextlib
+from test.support import EnvironmentVarGuard
 
 import factory
 import mock
 from django.test import SimpleTestCase
 from django.test import TestCase
 from facility_profile.models import Facility
+from facility_profile.models import InteractionLog
 from facility_profile.models import MyUser
 from facility_profile.models import SummaryLog
-from test.support import EnvironmentVarGuard
 
 from ..helpers import serialized_facility_factory
 from ..helpers import TestSessionContext
@@ -428,10 +429,12 @@ class DeserializationFromStoreIntoAppTestCase(TestCase):
         self.assertFalse(Facility.objects.filter(id=st.id).exists())
 
     def test_broken_fk_leaves_store_dirty_bit(self):
-        serialized = """{"user_id": "40de9a3fded95d7198f200c78e559353", "id": "bd205b5ee5bc42da85925d24c61341a8"}"""
+        log_id = uuid.uuid4().hex
+        serialized = json.dumps({"user_id": "40de9a3fded95d7198f200c78e559353", "id": log_id})
         st = StoreModelFacilityFactory(
-            id=uuid.uuid4().hex, serialized=serialized, model_name="contentsummarylog"
+            id=log_id, serialized=serialized, model_name="contentsummarylog"
         )
+        self.assertTrue(st.dirty_bit)
         self.mc.deserialize_from_store()
         st.refresh_from_db()
         self.assertTrue(st.dirty_bit)
@@ -614,6 +617,12 @@ class SelfReferentialFKDeserializationTestCase(TestCase):
         self.assertTrue(new_child.dirty_bit)
         self.assertIn("exist", new_child.deserialization_error)
 
+
+class ForeignKeyDeserializationTestCase(TestCase):
+    def setUp(self):
+        (self.current_id, _) = InstanceIDModel.get_or_create_current_instance()
+        self.mc = MorangoProfileController("facilitydata")
+
     def test_deserialization_of_model_with_missing_foreignkey_referent(self):
 
         user = MyUser.objects.create(username="penguin")
@@ -632,7 +641,68 @@ class SelfReferentialFKDeserializationTestCase(TestCase):
 
         new_log.refresh_from_db()
         self.assertTrue(new_log.dirty_bit)
-        self.assertIn("exist", new_log.deserialization_error)
+        self.assertIn("my user instance with id '{}'".format(data["user_id"]), new_log.deserialization_error)
+
+    def test_deserialization_of_model_with_disallowed_null_foreignkey(self):
+
+        user = MyUser.objects.create(username="penguin")
+        log = SummaryLog.objects.create(user=user)
+        self.mc.serialize_into_store()
+
+        new_log = Store.objects.get(id=log.id)
+        data = json.loads(new_log.serialized)
+        new_log.id = data["id"] = "f" * 32
+        data["user_id"] = None
+        new_log.serialized = json.dumps(data)
+        new_log.dirty_bit = True
+        new_log.save()
+
+        self.mc.deserialize_from_store()
+
+        new_log.refresh_from_db()
+        self.assertTrue(new_log.dirty_bit)
+        self.assertIn("cannot be null", new_log.deserialization_error)
+
+    def test_deserialization_of_model_with_allowed_null_foreignkey(self):
+
+        user = MyUser.objects.create(username="penguin")
+        log = InteractionLog.objects.create(user=user)
+        self.mc.serialize_into_store()
+
+        new_log = Store.objects.get(id=log.id)
+        data = json.loads(new_log.serialized)
+        new_log.id = data["id"] = "f" * 32
+        data["user_id"] = None
+        new_log.serialized = json.dumps(data)
+        new_log.dirty_bit = True
+        new_log.save()
+
+        self.mc.deserialize_from_store()
+
+        new_log.refresh_from_db()
+        self.assertFalse(new_log.dirty_bit)
+        self.assertTrue(new_log.deserialization_error == "")
+        self.assertTrue(InteractionLog.objects.filter(id=new_log.id).exists())
+
+    def test_deserialization_of_model_with_valid_foreignkey_referent(self):
+
+        user = MyUser.objects.create(username="penguin")
+        log = SummaryLog.objects.create(user=user)
+        self.mc.serialize_into_store()
+
+        new_log = Store.objects.get(id=log.id)
+        data = json.loads(new_log.serialized)
+        new_log.id = data["id"] = "f" * 32
+        new_log.serialized = json.dumps(data)
+        new_log.dirty_bit = True
+        new_log.save()
+
+        self.mc.deserialize_from_store()
+
+        new_log.refresh_from_db()
+        self.assertFalse(new_log.dirty_bit)
+        self.assertTrue(new_log.deserialization_error == "")
+        self.assertTrue(SummaryLog.objects.filter(id=new_log.id).exists())
 
 
 class SessionControllerTestCase(SimpleTestCase):